requirements.txt
python-dotenv
pydantic
email-validator
uvicorn[standard]
fastapi
sqlalchemy
httpx
pytest
pytest-dependency
argon2-cffi
python-jose[cryptography]
python-multipart
fuzzywuzzy
python-Levenshtein
